app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

weaviate-client==3.26.7

sentence-transformers==2.7.0
transformers==4.41.2
torch>=2.2.0

requests==2.32.3

pydantic==2.7.1